RF: Move add_readme() out of plugins
Towards a fix for #4519
